+++ TCAT/Controllers/RouteDetail+DrawerViewController.swift
@@ -20,27 +20,22 @@ struct RouteDetailCellSize {
 
 class RouteDetailDrawerViewController: UIViewController {
 
+    var safeAreaCover: UIView?
     var summaryView = SummaryView()
+    var tableView: UITableView!
 
-    private var safeAreaCover: UIView?
-    private var tableView: UITableView!
+    var directions: [Direction] = []
+    var justLoaded: Bool = true
+    var ongoing: Bool = false
+    var selectedDirection: Direction?
+    var visible: Bool = false
 
     /// Number of seconds to wait before auto-refreshing bus delay network call.
     private var busDelayNetworkRefreshRate: Double = 10
     private var busDelayNetworkTimer: Timer?
-    private var directions: [Direction] = []
-    private var justLoaded: Bool = true
-    private let main = UIScreen.main.bounds
-    private var ongoing: Bool = false
-    private var visible: Bool = false
-
-    private var selectedDirection: Direction?
-
-    /// Dictionary that maps the original indexPath.row to the number of cells that specific row added
-    var expandedCellDict: [Int: Int] = [:]
-
     private let chevronFlipDurationTime = 0.25
-
+    /// Dictionary that maps the original indexPath.row to the number of cells that specific row added
+    private var expandedCellDict: [Int: Int] = [:]
     private let networking: Networking = URLSession.shared.request
     private var route: Route!
 
@@ -110,7 +105,7 @@ class RouteDetailDrawerViewController: UIViewController {
         // Create Detail Table View
         tableView = UITableView()
         tableView.frame.origin = CGPoint(x: 0, y: summaryView.frame.height)
-        tableView.frame.size = CGSize(width: main.width, height: main.height - summaryView.frame.height)
+        tableView.frame.size = CGSize(width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height - summaryView.frame.height)
         tableView.bounces = false
         tableView.estimatedRowHeight = RouteDetailCellSize.smallHeight
         tableView.rowHeight = UITableView.automaticDimension
@@ -144,7 +139,7 @@ class RouteDetailDrawerViewController: UIViewController {
     func initializeCover() {
         if #available(iOS 11.0, *) {
             let bottom = UIApplication.shared.keyWindow?.rootViewController?.view.safeAreaInsets.bottom ?? 34
-            safeAreaCover = UIView(frame: CGRect(x: 0, y: summaryView.frame.height, width: main.width, height: bottom))
+            safeAreaCover = UIView(frame: CGRect(x: 0, y: summaryView.frame.height, width: UIScreen.main.bounds.width, height: bottom))
             safeAreaCover!.backgroundColor = Colors.backgroundWash
             safeAreaCover!.alpha = 0
             view.addSubview(safeAreaCover!)
@@ -266,223 +261,5 @@ class RouteDetailDrawerViewController: UIViewController {
 
         busStops = []
         indexPathArray = []
-    }
-}
-
-// MARK: Gesture Recognizers and Interaction-Related Functions
-extension RouteDetailDrawerViewController: UIGestureRecognizerDelegate {
-    /** Animate detailTableView depending on context, centering map */
-    @objc func summaryTapped(_ sender: UITapGestureRecognizer? = nil) {
-
-        if let drawer = self.parent as? RouteDetailViewController {
-            switch drawer.drawerPosition {
-            case .collapsed, .partiallyRevealed:
-                if selectedDirection != nil {
-                    drawer.setDrawerPosition(position: .collapsed, animated: true)
-                } else {
-                    drawer.setDrawerPosition(position: .open, animated: true)
-                }
-            case .open:
-                drawer.setDrawerPosition(position: .collapsed, animated: true)
-            default: break
-            }
-        }
-    }
-
-}
-
-extension RouteDetailDrawerViewController: LargeDetailTableViewDelegate {
-
-    func collapseCells(on cell: UITableViewCell) {
-        toggleCellExpansion(at: tableView.indexPath(for: cell))
-    }
-
-    func expandCells(on cell: UITableViewCell) {
-        if justLoaded { summaryTapped() }
-
-        toggleCellExpansion(at: tableView.indexPath(for: cell))
-
-        tableView.layoutIfNeeded()
-        tableView.layoutSubviews()
-    }
-}
-
-extension RouteDetailDrawerViewController: PulleyDrawerViewControllerDelegate {
-    func collapsedDrawerHeight(bottomSafeArea: CGFloat) -> CGFloat {
-        return bottomSafeArea + summaryView.frame.height
-    }
-
-    func partialRevealDrawerHeight(bottomSafeArea: CGFloat) -> CGFloat {
-        return main.height / 2
-    }
-
-    func drawerPositionDidChange(drawer: PulleyViewController, bottomSafeArea: CGFloat) {
-
-        justLoaded = false
-        // Center map on drawer change
-        switch drawer.drawerPosition {
-        case .collapsed, .partiallyRevealed:
-            guard let contentViewController = drawer.primaryContentViewController as? RouteDetailContentViewController
-                else { return }
-            if let direction = selectedDirection {
-                if direction.type == .walk {
-                    contentViewController.centerMap(on: direction, isOverviewOfPath: true)
-                } else {
-                    contentViewController.centerMap(on: direction)
-                }
-                selectedDirection = nil
-            } else {
-                contentViewController.centerMapOnOverview(drawerPreviewing: drawer.drawerPosition == .partiallyRevealed)
-            }
-        default: break
-        }
-    }
-
-    func drawerChangedDistanceFromBottom(drawer: PulleyViewController, distance: CGFloat, bottomSafeArea: CGFloat) {
-
-        // Manage cover view hiding drawer when collapsed
-        if distance - bottomSafeArea == summaryView.frame.height {
-            safeAreaCover?.alpha = 1.0
-            visible = true
-        } else {
-            if !ongoing && visible {
-                UIView.animate(withDuration: 0.25, animations: {
-                    self.safeAreaCover?.alpha = 0.0
-                    self.visible = false
-                }, completion: { _ in
-                    self.ongoing = false
-                })
-            }
-        }
-    }
-
-    func supportedDrawerPositions() -> [PulleyPosition] {
-        return [.collapsed, .partiallyRevealed, .open]
-    }
-}
-
-extension RouteDetailDrawerViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return directions.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-
-        let direction = directions[indexPath.row]
-        let isBusStopCell = direction.type == .walk && direction.startLocation.latitude == 0.0
-        let cellWidth: CGFloat = RouteDetailCellSize.regularWidth
-
-        /// Formatting, including selectionStyle, and seperator line fixes
-        func format(_ cell: UITableViewCell) -> UITableViewCell {
-            cell.selectionStyle = .none
-            if indexPath.row == directions.count - 1 {
-                // Remove seperator at end of table
-                cell.layoutMargins = UIEdgeInsets(top: 0, left: main.width, bottom: 0, right: 0)
-            }
-            return cell
-        }
-
-        if isBusStopCell {
-            let cell = tableView.dequeueReusableCell(withIdentifier: Constants.Cells.busStopDetailCellIdentifier) as! BusStopTableViewCell
-            cell.setCell(direction.name)
-            cell.layoutMargins = UIEdgeInsets(top: 0, left: cellWidth + 20, bottom: 0, right: 0)
-            return format(cell)
-        } else if direction.type == .walk || direction.type == .arrive {
-            let cell = tableView.dequeueReusableCell(withIdentifier: Constants.Cells.smallDetailCellIdentifier, for: indexPath) as! SmallDetailTableViewCell
-            cell.setCell(direction,
-                         firstStep: indexPath.row == 0,
-                         lastStep: indexPath.row == directions.count - 1)
-            cell.layoutMargins = UIEdgeInsets(top: 0, left: cellWidth, bottom: 0, right: 0)
-            return format(cell)
-        } else {
-            let cell = tableView.dequeueReusableCell(withIdentifier: Constants.Cells.largeDetailCellIdentifier) as! LargeDetailTableViewCell
-            cell.setCell(direction, indexPath: indexPath)
-            cell.delegate = self
-            cell.layoutMargins = UIEdgeInsets(top: 0, left: cellWidth, bottom: 0, right: 0)
-            return format(cell)
-        }
-
-    }
-}
-
-extension RouteDetailDrawerViewController: UITableViewDelegate {
-
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-
-        let direction = directions[indexPath.row]
-
-        if direction.type == .depart || direction.type == .transfer {
-            let cell = tableView.dequeueReusableCell(withIdentifier: Constants.Cells.largeDetailCellIdentifier) as? LargeDetailTableViewCell
-            cell?.setCell(direction, indexPath: indexPath)
-            return cell?.height() ?? RouteDetailCellSize.largeHeight
-        } else {
-            return RouteDetailCellSize.smallHeight
-        }
-
-    }
-
-    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-
-        // Empty Footer
-
-        let emptyFooterView = tableView.dequeueReusableHeaderFooterView(withIdentifier: Constants.Footers.emptyFooterView) ??
-            UITableViewHeaderFooterView(reuseIdentifier: Constants.Footers.emptyFooterView)
-
-        let lastCellIndexPath = IndexPath(row: tableView.numberOfRows(inSection: 0) - 1, section: 0)
-        var screenBottom = main.height
-        if #available(iOS 11.0, *) {
-            screenBottom -= view.safeAreaInsets.bottom
-        }
-
-        // Calculate height of space between last cell and the bottom of the screen, also accounting for summary
-        var footerHeight = screenBottom - (tableView.cellForRow(at: lastCellIndexPath)?.frame.maxY ?? screenBottom) - summaryView.frame.height
-        footerHeight = expandedCell != nil ? 0 : footerHeight
-
-        emptyFooterView.frame.size = CGSize(width: view.frame.width, height: footerHeight)
-        emptyFooterView.contentView.backgroundColor = Colors.white
-        emptyFooterView.layoutIfNeeded()
-
-        // Create Footer for No Data from Live Tracking Footer, if needed
-
-        guard
-            let drawer = self.parent as? RouteDetailViewController,
-            let contentViewController = drawer.primaryContentViewController as? RouteDetailContentViewController
-            else {
-                return emptyFooterView
-        }
-
-        var message: String?
-
-        if !contentViewController.noDataRouteList.isEmpty {
-            if contentViewController.noDataRouteList.count > 1 {
-                message = Constants.Banner.noLiveTrackingForRoutes
-            } else {
-                let routeNumber = contentViewController.noDataRouteList.first!
-                message = Constants.Banner.noLiveTrackingForRoute + " " + "\(routeNumber)."
-            }
-        } else {
-            message = nil
-        }
-
-        if let message = message {
-            let phraseLabelFooterView = tableView.dequeueReusableHeaderFooterView(withIdentifier: Constants.Footers.phraseLabelFooterView)
-                as? PhraseLabelFooterView ?? PhraseLabelFooterView(reuseIdentifier: Constants.Footers.phraseLabelFooterView)
-            phraseLabelFooterView.setView(with: message)
-            return phraseLabelFooterView
-        }
-
-        return emptyFooterView
-
-    }
-
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-
-        let direction = directions[indexPath.row]
-
-        selectedDirection = direction
-
-        if let drawer = self.parent as? RouteDetailViewController {
-            drawer.setDrawerPosition(position: .collapsed, animated: true)
-        }
     }
 }
